refactor: Introduce AbstractFetchAccessEgress and use in TransitRouter
