+++ application/src/main/java/org/opentripplanner/routing/algorithm/raptoradapter/router/TransitRouter.java
@@ -68,6 +68,7 @@ public class TransitRouter {
   private final AccessEgressRouter accessEgressRouter;
   private final TransitServiceResolver transitServiceResolver;
   private final CarpoolingService carpoolingService;
+  private final AbstractFetchAccessEgress fetchAccessEgress;
 
   private TransitRouter(
     RouteRequest request,
@@ -90,6 +91,16 @@ public class TransitRouter {
     this.transitServiceResolver = new TransitServiceResolver(serverContext.transitService());
     this.accessEgressRouter = new AccessEgressRouter(this.transitServiceResolver);
     this.carpoolingService = carpoolingService;
+    this.fetchAccessEgress = new AbstractFetchAccessEgress(
+      request,
+      serverContext,
+      transitSearchTimeZero,
+      additionalSearchDays,
+      linkingContext,
+      accessEgressRouter,
+      transitServiceResolver,
+      carpoolingService
+    );
   }
 
   public static TransitRouterResult route(
@@ -263,111 +274,16 @@ public class TransitRouter {
 
   private Collection<? extends RoutingAccessEgress> fetchAccess() {
     debugTimingAggregator.startedAccessCalculating();
-    var list = fetchAccessEgresses(ACCESS);
+    var list = fetchAccessEgress.fetchAccess();
     debugTimingAggregator.finishedAccessCalculating();
     return list;
   }
 
   private Collection<? extends RoutingAccessEgress> fetchEgress() {
     debugTimingAggregator.startedEgressCalculating();
-    var list = fetchAccessEgresses(EGRESS);
+    var list = fetchAccessEgress.fetchEgress();
     debugTimingAggregator.finishedEgressCalculating();
     return list;
-  }
-
-  private Collection<? extends RoutingAccessEgress> fetchAccessEgresses(AccessEgressType type) {
-    var streetRequest = type.isAccess() ? request.journey().access() : request.journey().egress();
-    StreetMode mode = streetRequest.mode();
-
-    // Prepare access/egress lists
-    var accessBuilder = request.copyOf();
-
-    if (type.isAccess()) {
-      accessBuilder.withPreferences(p -> {
-        p.withBike(b -> b.withRental(r -> r.withAllowArrivingInRentedVehicleAtDestination(false)));
-        p.withCar(c -> c.withRental(r -> r.withAllowArrivingInRentedVehicleAtDestination(false)));
-        p.withScooter(s ->
-          s.withRental(r -> r.withAllowArrivingInRentedVehicleAtDestination(false))
-        );
-      });
-    }
-
-    var accessRequest = accessBuilder.buildRequest();
-
-    var accessEgressPreferences = accessRequest.preferences().street().accessEgress();
-
-    Duration durationLimit = accessEgressPreferences.maxDuration().valueOf(mode);
-    int stopCountLimit = accessEgressPreferences.maxStopCountLimit().limitForMode(mode);
-
-    var nearbyStops = accessEgressRouter.findAccessEgresses(
-      accessRequest,
-      streetRequest.mode(),
-      serverContext.listExtensionRequestContexts(accessRequest),
-      type,
-      durationLimit,
-      stopCountLimit,
-      linkingContext
-    );
-    var accessEgresses = AccessEgressMapper.mapNearbyStops(nearbyStops, type);
-    accessEgresses = timeshiftRideHailing(streetRequest, type, accessEgresses);
-
-    var results = new ArrayList<>(accessEgresses);
-
-    // Special handling of flex accesses
-    if (OTPFeature.FlexRouting.isOn() && mode == StreetMode.FLEXIBLE) {
-      var flexAccessList = FlexAccessEgressRouter.routeAccessEgress(
-        accessRequest,
-        accessEgressRouter,
-        serverContext,
-        additionalSearchDays,
-        serverContext.flexParameters(),
-        serverContext.listExtensionRequestContexts(accessRequest),
-        type,
-        linkingContext
-      );
-
-      results.addAll(AccessEgressMapper.mapFlexAccessEgresses(flexAccessList, type));
-    }
-
-    if (OTPFeature.CarPooling.isOn() && mode == StreetMode.CARPOOL) {
-      var carpoolAccessEgressList = carpoolingService.routeAccessEgress(
-        accessRequest,
-        streetRequest,
-        type,
-        transitServiceResolver,
-        linkingContext,
-        transitSearchTimeZero
-      );
-      results.addAll(carpoolAccessEgressList);
-    }
-
-    return results;
-  }
-
-  /**
-   * Given a list of {@code results} shift the access ones that contain driving so that they only
-   * start at the time when the ride hailing vehicle can actually be there to pick up passengers.
-   * <p>
-   * If there are accesses/egresses with only walking, then they remain unchanged.
-   * <p>
-   * This method is a good candidate to be moved to the access/egress filter chain when that has
-   * been added.
-   */
-  private List<RoutingAccessEgress> timeshiftRideHailing(
-    StreetRequest streetRequest,
-    AccessEgressType type,
-    List<RoutingAccessEgress> accessEgressList
-  ) {
-    if (streetRequest.mode() != StreetMode.CAR_HAILING) {
-      return accessEgressList;
-    }
-    return RideHailingAccessShifter.shiftAccesses(
-      type.isAccess(),
-      accessEgressList,
-      serverContext.rideHailingServices(),
-      request,
-      Instant.now()
-    );
   }
 
   private RaptorRoutingRequestTransitData createRequestTransitDataProvider(
